Support full-disk encryption for Hygon CSV/CSV2/CSV3 guest
When we place grub FV package into OVMF.fd to support full-disk
encryption, we need provide grub components as the build env.

===
Prerequisite:

1. If the grub in your system has already support efisecret module, then
you should run cmds as follows to support build OVMF.fd.
  $ cd /usr
  $ sudo tar --transform='s/^/grub\// -zcf grub.tar.gz lib/grub/
  $ sudo mv grub.tar.gz /opt
  $ cd /opt/
  $ sudo tar -xzf grub.tar.gz

2. If the grub in your system does not support efisecret module, then we
can clone the grub repository and checkout to commit 578c95298 ("
kern: Add lockdown support"), and backport the following patch series:
  https://lists.gnu.org/archive/html/grub-devel/2020-12/msg00257.html
  https://lists.gnu.org/archive/html/grub-devel/2020-12/msg00258.html
  https://lists.gnu.org/archive/html/grub-devel/2020-12/msg00259.html
  https://lists.gnu.org/archive/html/grub-devel/2020-12/msg00260.html
After we have prepared for the grub codes, we should build it and
install the grub components to /opt/grub. The steps is shown as below:
  $ ./bootstrap
  $ ./autogen.sh
  $ ./configure --target=x86_64 --with-platform=efi -prefix=/grub
  $ make -j$(getconf _NPROCESSORS_ONLN)
  $ sudo make install DESTDIR=/opt/
  $ cd /opt/
  $ sudo tar -zcf grub.tar.gz grub

Note: the grub.tar.gz metioned above will be used when enabling
full-disk encryption in the guest.

===
Build OVMF.fd:

When we build the OVMF.fd which support full-disk encryption, we should
delete stale grub.efi in the source tree:
  $ rm OvmfPkg/AmdSev/Grub/grub.efi
And specify the dsc file:
  $ OvmfPkg/build.sh ... -a X64 -p OvmfPkg/AmdSev/AmdSevX64.dsc ...

Signed-off-by: hanliyang <hanliyang@hygon.cn>