debian/rules: Add build-ovmf-csv target to build OVMF.CSV.fd
===
Prerequisite:

See the `Prerequisite 1.` in the previous commit's message.

===
Build OVMF.CSV.fd:

Actually, the OVMF.CSV.fd mentioned here is the OVMF.fd mentioned in the
previous commit's message. In order build the OVMF.fd and save it, we
rename it to OVMF.CSV.fd.

When we build the OVMF.CSV.fd which support full-disk encryption, we
should delete stale grub.efi in the source tree:
  $ rm OvmfPkg/AmdSev/Grub/grub.efi
And specify build target:
  $ dpkg-buildpackage -us -uc -b -T build-ovmf-csv

After completion of the build process, the OVMF.CSV.fd will be saved to
debian/ovmf-install/OVMF.CSV.fd.

Signed-off-by: hanliyang <hanliyang@hygon.cn>